rework src to allow for various module patterns
